Extract Irdeto-specific construction
